Conflicts:
	recipes/secure_hbase_tm-3.rb

+++ bin/hbase-ec2-init-remote.sh
@@ -1,8 +1,26 @@
 #!/usr/bin/env bash
 
+#
+# Copyright 2010 The Apache Software Foundation
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 set -x
 export JAVA_HOME=/usr/local/jdk1.6.0_20
-
 ln -s $JAVA_HOME /usr/local/jdk
 
 # Script that is run on each EC2 instance on boot. It is passed in the EC2 user
@@ -77,21 +95,18 @@ add_principal -pw $kadmpass hadoop/admin
 add_principal -pw had00p hclient
 quit
 EOF
-
 }
 
-
 # up file-max
-sysctl -w fs.file-max=32768
+sysctl -w fs.file-max=65535
 
 # up ulimits
-echo "root soft nofile 32768" >> /etc/security/limits.conf
-echo "root hard nofile 32768" >> /etc/security/limits.conf
-echo "hadoop soft nofile 32768" >> /etc/security/limits.conf
-echo "hadoop hard nofile 32768" >> /etc/security/limits.conf
+echo "root soft nofile 65535" >> /etc/security/limits.conf
+echo "root hard nofile 65535" >> /etc/security/limits.conf
+ulimit -n 65535
 
 # up epoll limits; ok if this fails, only valid for kernels 2.6.27+
-sysctl -w fs.epoll.max_user_instances=32768 > /dev/null 2>&1
+sysctl -w fs.epoll.max_user_instances=65535 > /dev/null 2>&1
 
 [ ! -f /etc/hosts ] &&  echo "127.0.0.1 localhost" > /etc/hosts
 echo "$HOST_IP $HOSTNAME" >> /etc/hosts
@@ -111,13 +126,12 @@ fi
 
 # Security setup
 # all servers need krb5 libraries
-yum -y install krb5-libs jakarta-commons-daemon-jsvc
-ln -s /usr/bin/jsvc $HADOOP_HOME/bin
+yum -y install jakarta-commons-daemon-jsvc
+[ -f $HADOOP_HOME/bin/jsvc ] || ln -s /usr/bin/jsvc $HADOOP_HOME/bin
 adduser hadoop
 groupadd supergroup
 adduser -G supergroup hbase
 if [ "$IS_MASTER" = "true" ]; then
-  yum -y install krb5-server
   cat > /var/kerberos/krb5kdc/kadm5.acl <<EOF
 */admin@HADOOP.LOCALDOMAIN    *
 EOF
@@ -186,7 +200,7 @@ cat > /etc/krb5.conf <<EOF
 	krb4_get_tickets = false
 EOF
 
-# TODO: generate these from pwgen and pass throug
+# TODO: generate these from pwgen and pass through.
 KDC_MASTER_PASS="EiSei0Da"
 KDC_ADMIN_PASS="Chohpet6"
 
@@ -200,14 +214,18 @@ keytab="$HADOOP_HOME/conf/nn.keytab"
 add_client "hadoop/admin" $KDC_ADMIN_PASS $keytab $HOSTNAME
 chown hadoop:hadoop $keytab
 
+if [ "$IS_MASTER" = "true" ]; then
+  cd /usr/local/hadoop-*; kinit -k -t conf/nn.keytab hadoop/$HOSTNAME
+fi
+
 # Ganglia
 
 if [ "$IS_MASTER" = "true" ]; then
   sed -i -e "s|\( *mcast_join *=.*\)|#\1|" \
-      -e "s|\( *bind *=.*\)|#\1|" \
-      -e "s|\( *mute *=.*\)|  mute = yes|" \
-      -e "s|\( *location *=.*\)|  location = \"master-node\"|" \
-      /etc/gmond.conf
+         -e "s|\( *bind *=.*\)|#\1|" \
+         -e "s|\( *mute *=.*\)|  mute = yes|" \
+         -e "s|\( *location *=.*\)|  location = \"master-node\"|" \
+         /etc/gmond.conf
   mkdir -p /mnt/ganglia/rrds
   chown -R ganglia:ganglia /mnt/ganglia/rrds
   rm -rf /var/lib/ganglia; cd /var/lib; ln -s /mnt/ganglia ganglia; cd
@@ -216,9 +234,9 @@ if [ "$IS_MASTER" = "true" ]; then
   apachectl start
 else
   sed -i -e "s|\( *mcast_join *=.*\)|#\1|" \
-      -e "s|\( *bind *=.*\)|#\1|" \
-      -e "s|\(udp_send_channel {\)|\1\n  host=$MASTER_HOST|" \
-      /etc/gmond.conf
+         -e "s|\( *bind *=.*\)|#\1|" \
+         -e "s|\(udp_send_channel {\)|\1\n  host=$MASTER_HOST|" \
+         /etc/gmond.conf
   service gmond start
 fi
 
@@ -416,6 +434,10 @@ cat > $HADOOP_HOME/conf/mapred-site.xml <<EOF
   <value>true</value>
 </property>
 <property>
+  <name>mapreduce.cluster.job-authorization-enabled</name>
+  <value>true</value>
+</property>
+<property>
   <name>mapreduce.job.acl-modify-job</name>
   <value></value>
 </property>
@@ -447,7 +469,7 @@ export HADOOP_OPTS="$HADOOP_OPTS -XX:+UseCompressedOops"
 EOF
 # Update classpath to include HBase jars and config
 cat >> $HADOOP_HOME/conf/hadoop-env.sh <<EOF
-export HADOOP_CLASSPATH="$HBASE_HOME/hbase-${HBASE_VERSION}.jar:$HBASE_HOME/lib/AgileJSON-2009-03-30.jar:$HBASE_HOME/lib/json.jar:$HBASE_HOME/lib/zookeeper-3.3.0.jar:$HBASE_HOME/conf"
+HADOOP_CLASSPATH="$HBASE_HOME/hbase-${HBASE_VERSION}.jar:$HBASE_HOME/lib/zookeeper-3.3.1.jar:$HBASE_HOME/conf"
 EOF
 # Configure Hadoop for Ganglia
 cat > $HADOOP_HOME/conf/hadoop-metrics.properties <<EOF
@@ -502,20 +524,8 @@ cat > $HBASE_HOME/conf/hbase-site.xml <<EOF
   <value>100</value>
 </property>
 <property>
-  <name>hfile.block.cache.size</name>
-  <value>0.3</value>
-</property>
-<property>
-  <name>hbase.regionserver.global.memstore.upperLimit</name>
-  <value>0.3</value>
-</property>
-<property>
-  <name>hbase.regionserver.global.memstore.lowerLimit</name>
-  <value>0.25</value>
-</property>
-<property>
   <name>hbase.hregion.memstore.block.multiplier</name>
-  <value>4</value>
+  <value>3</value>
 </property>
 <property>
   <name>hbase.hstore.blockingStoreFiles</name>
@@ -586,9 +596,8 @@ ln -s $HADOOP_HOME/conf/mapred-site.xml $HBASE_HOME/conf/
 # Override JVM options
 cat >> $HBASE_HOME/conf/hbase-env.sh <<EOF
 export JAVA_HOME=/usr/local/jdk
-export HBASE_LOG_DIR=/mnt/hbase/logs
-export HBASE_MASTER_OPTS="-Xmx1000m -XX:+UseCompressedOops -XX:+UseConcMarkSweepGC -XX:+DoEscapeAnalysis -XX:+AggressiveOpts -verbose:gc -XX:+PrintGCDetails -XX:+PrintGCDateStamps -Xloggc:/mnt/hbase/logs/hbase-master-gc.log"
-export HBASE_REGIONSERVER_OPTS="-Xmx4000m -XX:+UseCompressedOops -XX:+UseConcMarkSweepGC -XX:CMSInitiatingOccupancyFraction=88 -XX:NewSize=128m -XX:MaxNewSize=128m -XX:+DoEscapeAnalysis -XX:+AggressiveOpts -verbose:gc -XX:+PrintGCDetails -XX:+PrintGCDateStamps -Xloggc:/mnt/hbase/logs/hbase-regionserver-gc.log"
+export HBASE_MASTER_OPTS="-Xmx1000m -XX:+UseConcMarkSweepGC -XX:NewSize=128m -XX:MaxNewSize=128m -XX:+AggressiveOpts -verbose:gc -XX:+PrintGCDetails -XX:+PrintGCTimeStamps -Xloggc:/mnt/hbase/logs/hbase-master-gc.log"
+export HBASE_REGIONSERVER_OPTS="-Xmx2000m -XX:+UseConcMarkSweepGC -XX:CMSInitiatingOccupancyFraction=88 -XX:NewSize=128m -XX:MaxNewSize=128m -XX:+AggressiveOpts -verbose:gc -XX:+PrintGCDetails -XX:+PrintGCTimeStamps -Xloggc:/mnt/hbase/logs/hbase-regionserver-gc.log"
 EOF
 # Configure log4j
 sed -i -e 's/hadoop.hbase=DEBUG/hadoop.hbase=INFO/g' \
@@ -615,12 +624,16 @@ if [ "$IS_MASTER" = "true" ]; then
   [ ! -e /mnt/hadoop/dfs/name ] && "$HADOOP_HOME"/bin/hadoop namenode -format
   "$HADOOP_HOME"/bin/hadoop-daemon.sh start namenode
   "$HADOOP_HOME"/bin/hadoop-daemon.sh start jobtracker
-  # "$HBASE_HOME"/bin/hbase-daemon.sh start master
+
+  #<must be done after hadoop startup, and before hbase startup>
+  "$HADOOP_HOME"/bin/hadoop fs -mkdir /hbase
+  "$HADOOP_HOME"/bin/hadoop fs -chown hbase /hbase
+  #</must be done after hadoop startup, and before hbase startup>
+
 else
     if [ "$IS_AUX" != "true" ]; then
 	"$HADOOP_HOME"/bin/hadoop-daemon.sh start datanode
-  # "$HBASE_HOME"/bin/hbase-daemon.sh start regionserver
-	"$HADOOP_HOME"/bin/hadoop-daemon.sh start tasktracker
+        "$HADOOP_HOME"/bin/hadoop-daemon.sh start tasktracker
     fi
 fi
